+++ packages/lumx-react/src/components/side-navigation/SideNavigationItem.tsx
@@ -2,6 +2,7 @@ import React, { Children, forwardRef, ReactNode } from 'react';
 
 import classNames from 'classnames';
 import isEmpty from 'lodash/isEmpty';
+import { uid } from 'uid';
 
 import { mdiChevronDown, mdiChevronUp } from '@lumx/icons';
 
@@ -94,6 +95,14 @@ export const SideNavigationItem: Comp<SideNavigationItemProps, HTMLLIElement> = 
     const shouldSplitActions = Boolean(onActionClick);
     const showChildren = hasContent && isOpen;
 
+    const contentId = React.useMemo(uid, []);
+    const ariaProps: any = {};
+    if (hasContent) {
+        ariaProps['aria-expanded'] = !!showChildren;
+        // Associate with content ID only if in DOM (shown or hidden and not unmounted)
+        ariaProps['aria-controls'] = showChildren || closeMode === 'hide' ? contentId : undefined;
+    }
+
     return (
         <li
             ref={ref}
@@ -129,6 +138,7 @@ export const SideNavigationItem: Comp<SideNavigationItemProps, HTMLLIElement> = 
                         size={Size.m}
                         emphasis={Emphasis.low}
                         onClick={onActionClick}
+                        {...ariaProps}
                     />
                 </div>
             ) : (
@@ -140,6 +150,7 @@ export const SideNavigationItem: Comp<SideNavigationItemProps, HTMLLIElement> = 
                         tabIndex: 0,
                         onClick,
                         onKeyDown: onClick ? onEnterPressed(onClick as Callback) : undefined,
+                        ...ariaProps,
                     },
                     icon && <Icon className={`${CLASSNAME}__icon`} icon={icon} size={Size.xs} />,
                     <span>{label}</span>,
@@ -153,7 +164,11 @@ export const SideNavigationItem: Comp<SideNavigationItemProps, HTMLLIElement> = 
                 )
             )}
 
-            {(closeMode === 'hide' || showChildren) && <ul className={`${CLASSNAME}__children`}>{content}</ul>}
+            {(closeMode === 'hide' || showChildren) && (
+                <ul className={`${CLASSNAME}__children`} id={contentId}>
+                    {content}
+                </ul>
+            )}
         </li>
     );
 });